+++ src/libertem/common/executor.py
@@ -1,11 +1,15 @@
 import queue
+import time
+import functools
 from typing import (
     Callable, Optional, Any, TYPE_CHECKING,
-    TypeVar,
+    TypeVar, Union
 )
 from collections.abc import Generator, Iterable
 from contextlib import contextmanager
 import multiprocessing as mp
+import threading
+import contextlib
 
 import cloudpickle
 from opentelemetry import trace
@@ -132,10 +136,12 @@ class JobExecutor:
     '''
     Interface to execute functions on workers.
     '''
-    def scale_up(self, n_workers: Optional[int]):
-        pass
-
-    def scale_down(self, n_workers: Optional[int]):
+    def scale(self, n_workers: Optional[int] = None):
+        """
+        Adjust the executor state to the given number of workers
+        If n_workers is None, return executor to its initial state
+        if not currently the case
+        """
         pass
 
     def run_function(self, fn: Callable[..., T], *args, **kwargs) -> T:
@@ -711,3 +717,68 @@ class NoopCommHandler(TaskCommHandler):
 
     def done(self):
         pass
+
+
+class SnoozeMixin:
+    def setup_snooze(self, snooze_timeout: Union[float, int]):
+        self._keep_alive = 0
+        self._last_activity = time.monotonic()
+        self.is_snoozing = False
+        self._snooze_lock = threading.Lock()
+        self._snooze_timeout = snooze_timeout
+        self._snooze_check_interval = min(
+            30.0,
+            self._snooze_timeout and (self._snooze_timeout * 0.1) or 30.0,
+        )
+        self._snooze_task = threading.Thread(
+            target=self._snooze_check_task,
+            daemon=True,
+        )
+        self._snooze_task.start()
+
+    def _update_last_activity(self):
+        self._last_activity = time.monotonic()
+
+    @contextlib.contextmanager
+    def in_use(self):
+        self._update_last_activity()
+        self._keep_alive += 1
+        try:
+            yield
+        finally:
+            self._keep_alive -= 1
+            self._update_last_activity()
+
+    def snooze(self):
+        if self._keep_alive > 0:
+            return
+        with self._snooze_lock:
+            self.scale(1)
+            self.is_snoozing = True
+
+    def unsnooze(self):
+        with self._snooze_lock:
+            self.scale()
+            self.is_snoozing = False
+
+    def _snooze_check_task(self):
+        """
+        Periodically check if we need to snooze the executor
+        """
+        while True:
+            time.sleep(self._snooze_check_interval)
+            if self.is_snoozing or self._keep_alive > 0:
+                continue
+            since_last_activity = time.monotonic() - self._last_activity
+            if since_last_activity > self._snooze_timeout:
+                self.snooze()
+
+    @staticmethod
+    def keep_alive(fn):
+
+        @functools.wraps(fn)
+        def wrapped(self, *args, **kwargs):
+            with self.in_use():
+                return fn(self, *args, **kwargs)
+
+        return wrapped
